STG Website
I have now completed the navbar, while sucessfully creating css files for each of the html files, and organised everything into an assets folder. I have linked all of these stylesheets together, along with all of the html pages share a css page. This page contains the navbar coding, for it will be the same in all of the pages.