STG Website 1.10
HUGE UPDATE. I worked very hard on this, including learning some Javascript. I added my photo (news) gallery at the bottom of the homepage, and made this page overall more responsive to all screens. I have made text to go along with each of the images in the gallery, which is able to be scrolled through with the use of arrows.